confirm delete

+++ index.js
@@ -81,15 +81,23 @@
       return true;
     }
 
-    $.ajax({
-      url: url,
-      type: 'delete',
-      data: data,
-      success: function(data) {
-        redirect_to(data.location);
+    var message = data["confirm"];
+    if (message) {
+      var sure = confirm(message);
+      if (sure) {
+        $.ajax({
+          url: url,
+          type: 'delete',
+          data: data,
+          success: function(data) {
+            redirect_to(data.location);
+          }
+        });
+        stopAll(e);
+      } else {
+        return true;
       }
-    });
-    stopAll(e);
+    }
   }
 
   // https://stackoverflow.com/questions/1184624/convert-form-data-to-javascript-object-with-jquery
